+++ backend/src/main/java/com/prograngers/backend/service/CommentService.java
@@ -8,10 +8,10 @@ import com.prograngers.backend.dto.comment.response.ShowMyCommentsResponse;
 import com.prograngers.backend.entity.comment.Comment;
 import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.solution.Solution;
-import com.prograngers.backend.exception.badrequest.DifferentSolutionException;
-import com.prograngers.backend.exception.badrequest.InvalidPageNumberException;
-import com.prograngers.backend.exception.badrequest.InvalidPageSizeException;
-import com.prograngers.backend.exception.badrequest.InvalidParentException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.DifferentSolutionException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.InvalidPageNumberException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.InvalidPageSizeException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.InvalidParentException;
 import com.prograngers.backend.exception.notfound.CommentAlreadyDeletedException;
 import com.prograngers.backend.exception.notfound.CommentNotFoundException;
 import com.prograngers.backend.exception.notfound.MemberNotFoundException;
@@ -22,7 +22,6 @@ import com.prograngers.backend.repository.member.MemberRepository;
 import com.prograngers.backend.repository.solution.SolutionRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
