+++ backend/src/main/java/com/prograngers/backend/service/ReviewService.java
@@ -11,10 +11,10 @@ import com.prograngers.backend.dto.review.response.ShowReviewsResponse;
 import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.review.Review;
 import com.prograngers.backend.entity.solution.Solution;
-import com.prograngers.backend.exception.badrequest.DifferentCodeLineNumberException;
-import com.prograngers.backend.exception.badrequest.DifferentSolutionException;
-import com.prograngers.backend.exception.badrequest.InvalidCodeLIneNumberException;
-import com.prograngers.backend.exception.badrequest.InvalidParentException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.DifferentCodeLineNumberException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.DifferentSolutionException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.InvalidCodeLIneNumberException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.InvalidParentException;
 import com.prograngers.backend.exception.notfound.MemberNotFoundException;
 import com.prograngers.backend.exception.notfound.ReviewAlreadyDeletedException;
 import com.prograngers.backend.exception.notfound.ReviewNotFoundException;
