+++ backend/src/test/java/com/prograngers/backend/service/SolutionServiceTest.java
@@ -48,7 +48,7 @@ import com.prograngers.backend.entity.solution.AlgorithmConstant;
 import com.prograngers.backend.entity.solution.DataStructureConstant;
 import com.prograngers.backend.entity.solution.LanguageConstant;
 import com.prograngers.backend.entity.solution.Solution;
-import com.prograngers.backend.exception.badrequest.PrivateSolutionException;
+import com.prograngers.backend.exception.badrequest.invalidvalue.PrivateSolutionException;
 import com.prograngers.backend.exception.notfound.ProblemLinkNotFoundException;
 import com.prograngers.backend.exception.unauthorization.MemberUnAuthorizedException;
 import com.prograngers.backend.repository.badge.BadgeRepository;
